Move test helper functions to new package
